refactor product entity and service: change price column type to float and implement product creation logic

+++ teslo-shop/src/products/entities/product.entity.ts
@@ -8,7 +8,7 @@ export class Product {
   @Column('text', { unique: true })
   title: string;
 
-  @Column('numeric')
+  @Column('float')
   price: number;
 
   @Column('text', { nullable: true })
